QD_GoogleAuth/requirements.txt
authlib==0.11
google-auth==1.6.3
google-auth-oauthlib==0.4.0
google-auth-httplib2==0.0.3
google-api-python-client==1.7.11
flask==1.1.1
requests==2.22.0
